added styles on menu page

+++ src/page/main/components/Menu/DataFrame.js
@@ -6,18 +6,26 @@ import { deleteDoc, doc } from "firebase/firestore/lite";
 import { db } from "../../../../config/adminFirebase";
 import toast from "react-hot-toast";
 import { useState } from "react";
+import { FaEdit, FaTrashAlt } from "react-icons/fa";
 
 const DataFrame = ({ data, title, rerender, setRerender }) => {
-  const [onDeleteBtn, setOnDeleteBtn] = useState(true);
+  const [onEditBtn, setOnEditBtn] = useState(true);
+
   const deleteMenuHandler = async (id) => {
-    await deleteDoc(doc(db, "menu", id))
-      .then(() => {
-        setRerender(!rerender);
-        toast.success("Menu deleted Successfully");
-        setOnDeleteBtn(false)
-      })
-      .catch((err) => toast.error("Menu delete failed!"));
+    const confirmed = window.confirm("Are you want to delete?");
+    if (confirmed) {
+      await deleteDoc(doc(db, "menu", id))
+        .then(() => {
+          setRerender(!rerender);
+          toast.success("Menu deleted Successfully");
+        })
+        .catch((err) => toast.error("Menu delete failed!"));
+    } else {
+      toast.error("Menu delete cancel!");
+    }
   };
+
+  const editMenuHandler = (id) => {};
 
   return (
     <div className="w-full border border-gray-300 flex flex-col">
@@ -34,64 +42,55 @@ const DataFrame = ({ data, title, rerender, setRerender }) => {
           </div>
         ))}
       </div>
-      {
-        data.map((t, index) => (
-          <div className="w-full" style={{ display: "flex", fontSize: 12 }}>
-            <div
-              className="w-24 border border-gray-100 p-3"
-              style={{ fontSize: extreSmallFont }}
-            >
-              {index + 1}
-            </div>
-            <div
-              className="flex-1 border border-gray-100 p-3"
-              style={{ fontSize: extreSmallFont }}
-            >
-              {t.foodName}
-            </div>
-            <div
-              className="flex-1 border border-gray-100 p-3"
-              style={{ fontSize: extreSmallFont }}
-            >
-              {t.category}
-            </div>
-            <div
-              className="flex-1 border border-gray-100 p-3"
-              style={{ fontSize: extreSmallFont }}
-            >
-              Open Recipe
-            </div>
-            <div className="flex-1 border border-gray-100 flex items-center justify-between">
-              <div className=" p-3" style={{ fontSize: extreSmallFont }}>
-                {t.price}
-              </div>
-
-              <div className="dropdown dropdown-left">
-                <button
-                onClick={()=> setOnDeleteBtn(true)}
-                  tabIndex={0}
-                  className="ml-auto bg-black rounded m-1 p-2"
-                >
-                  <IoChevronDown size={12} color="#fff" />
-                </button>
-                {onDeleteBtn && <ul
-                  tabIndex={0}
-                  className="dropdown-content menu p-2 shadow bg-base-100 rounded-box w-20"
-                >
-                  <li>
-                    <button
-                      onClick={() => deleteMenuHandler(t.id)}
-                      className="py-2 bg-red-700 hover:bg-red-800 text-white text-center"
-                    >
-                      Delete
-                    </button>
-                  </li>
-                </ul>}
-              </div>
-            </div>
+      {data.map((t, index) => (
+        <div className="w-full" style={{ display: "flex", fontSize: 12 }}>
+          <div
+            className="w-24 border border-gray-100 p-3"
+            style={{ fontSize: extreSmallFont }}
+          >
+            {index + 1}
           </div>
-        ))
-      }
+          <div
+            className="flex-1 border border-gray-100 p-3"
+            style={{ fontSize: extreSmallFont }}
+          >
+            {t.foodName}
+          </div>
+          <div
+            className="flex-1 border border-gray-100 p-3"
+            style={{ fontSize: extreSmallFont }}
+          >
+            {t.category}
+          </div>
+          <div
+            className="flex-1 border border-gray-100 p-3"
+            style={{ fontSize: extreSmallFont }}
+          >
+            Open Recipe
+          </div>
+          <div
+            className="flex-1 border border-gray-100 p-3"
+            style={{ fontSize: extreSmallFont }}
+          >
+            {t.price}
+          </div>
+
+          <div className="border border-gray-100 py-3 px-6 flex gap-4">
+            <button
+              onClick={() => editMenuHandler(t.id)}
+              className="duration-500 text-green-600 bg-white hover:bg-green-600 border border-green-600 hover:text-white py-2 px-2 rounded"
+            >
+              <FaEdit />
+            </button>
+            <button
+              onClick={() => deleteMenuHandler(t.id)}
+              className="duration-500 text-red-600 bg-white hover:bg-red-600 border border-red-600 hover:text-white py-2 px-2 rounded"
+            >
+              <FaTrashAlt />
+            </button>
+          </div>
+        </div>
+      ))}
     </div>
   );
 };
